+++ test/blackbox-tests/test-cases/install-dir/install-datadir.t
@@ -7,7 +7,7 @@
   > EOF
   $ dune build @install
   $ mkdir install datadir
-  $ dune install --dry-run --prefix ./install --datadir $PWD/datadir 2>&1 | grep datadir
+  $ dune install --dry-run --prefix ./install --datadir $PWD/datadir --display short 2>&1 | grep datadir
   Removing (if it exists) $TESTCASE_ROOT/datadir/foo/datafile
   Installing $TESTCASE_ROOT/datadir/foo/datafile
   Creating directory $TESTCASE_ROOT/datadir/foo