+++ test/blackbox-tests/test-cases/install-dir/install-docdir.t
@@ -7,7 +7,7 @@
   > EOF
   $ dune build @install
   $ mkdir install docdir
-  $ dune install --dry-run --prefix ./install --docdir $PWD/docdir 2>&1 | grep docdir
+  $ dune install --dry-run --prefix ./install --docdir $PWD/docdir --display short 2>&1 | grep docdir
   Removing (if it exists) $TESTCASE_ROOT/docdir/foo/docfile
   Installing $TESTCASE_ROOT/docdir/foo/docfile
   Creating directory $TESTCASE_ROOT/docdir/foo