+++ test/blackbox-tests/test-cases/install-multiple-contexts.t
@@ -27,12 +27,12 @@ Cannot install into a specific prefix with multiple contexts defined:
 
 One must pass a --context argument:
 
-  $ dune install --prefix _install/a --context a
+  $ dune install --prefix _install/a --context a --display short
   Installing _install/a/lib/a/META
   Installing _install/a/lib/a/dune-package
   Installing _install/a/share/a/hello
 
-  $ dune install --prefix _install/b --context b
+  $ dune install --prefix _install/b --context b --display short
   Installing _install/b/lib/a/META
   Installing _install/b/lib/a/dune-package
   Installing _install/b/share/a/hello