+++ test/blackbox-tests/test-cases/melange/basic-install.t
@@ -44,7 +44,7 @@ Test that we can install melange mode libraries
      (visibility public)
      (source (path Foo) (impl (path foo.ml))))))
 
-  $ dune install --prefix prefix
+  $ dune install --prefix prefix --display short
   Installing prefix/lib/foo/META
   Installing prefix/lib/foo/dune-package
   Installing prefix/lib/foo/foo.ml