+++ test/blackbox-tests/test-cases/melange/emit-installed-two-modes.t
@@ -21,7 +21,7 @@ Test dependency on installed package
   Entering directory 'a'
   Leaving directory 'a'
 
-  $ dune install --root a --prefix $PWD/prefix
+  $ dune install --root a --prefix $PWD/prefix --display short
   Installing $TESTCASE_ROOT/prefix/lib/a/META
   Installing $TESTCASE_ROOT/prefix/lib/a/a.cma
   Installing $TESTCASE_ROOT/prefix/lib/a/a.cmi