+++ test/blackbox-tests/test-cases/stdlib-compilation-depend-on-library.t
@@ -28,7 +28,7 @@ Test dependency on installed package
   Entering directory 'a'
   Leaving directory 'a'
 
-  $ dune install --root a --prefix $PWD/prefix
+  $ dune install --root a --prefix $PWD/prefix --display short
   Installing $TESTCASE_ROOT/prefix/lib/a/META
   Installing $TESTCASE_ROOT/prefix/lib/a/a.a
   Installing $TESTCASE_ROOT/prefix/lib/a/a.cma